this is my last correction

+++ trap.java
@@ -4,11 +4,11 @@ class trap{
 	private double f(double x){
 		return 1/(1+x*x);
 	}
-	trap(int l,int u,int n){
-	double h=(u-l);
+	trap(float l,float u,float n){
+	    double h=(u-l);
 		h/=n;
 		sum=f(l)+f(u);
-		for(int i=1;i<n;i++)
+		for(float i=1;i<n;i++)
 		{
 			k=l+i*h;
 			sum+=2*f(k);
