+++ trapezoidal.java
@@ -5,9 +5,9 @@ class trapezoidal{
 	public static void main(String[] args)
 	{
 		Scanner sc=new Scanner(System.in);
-		int l=sc.nextInt();
-		int u=sc.nextInt();
-		int n=sc.nextInt();
+		float l=sc.nextFloat();
+		float u=sc.nextFloat();
+		float n=sc.nextFloat();
 		
 		trap tp=new trap(l,u,n);
